Add --[no-]cache and --cache-dir options.
By default the caching is disabled.  When turned on with --cache,
the caching dir will be <output-dir>/.cache/ unless overridden
with --cache-dir option.

Previously the whole output dir got cleaned up, but now it can
contain a .cache dir, which we'll want to preserve.  So I created
OutputDir class to take care of output directory cleanup in a way
that preserves the .cache/ directory inside it.

Refs #446