Implement @experimental tag.
Placed it in the same group with @deprecated and @removed, so one
can also add a version number (since when the feature was added),
and a comment:

    @experimental 3.2 To try out new Google API.

Tweaked the styles of @deprecated and @removed too, so they all look
a bit better.  To allow for different message text in @experimental,
had to introduce @msg and @since instance variables to the DeprecatedTag
base class.

Fixes #239