Change @class_icon field to hold icon filename.
Copy the files over to class_icons/ dir like the member icons
are copied to member_icons/.  Additionally the @class tag now
defines the default icon.

Similarly also generate CSS for the icons.

This change only deals with class icons in class header (icons in
tree and search are to do).