Speed up options testing.
Reset the defaults every time the Options::Parser#parse method is
called.  This allows us to re-use the parser which is slow to instantiate.

Some tests still need to re-instatiate the parser every time - move
these to a separate suite, so only the tests that need it will run
slower.