Change URL separator character from hyphen to tilde
URLs for members in the generated documentation previously used the
hyphen character to separate the class name from the member ID, so the
website would be unable to parse URLs unambiguously. The separator is
now the tilde character.

This patch therefore adds support for class names which include hyphen
characters.